Bug 4564: Implement datastore restore from backup file
Added a singleton DatastoreSnapshotRestore class that looks for and reads a
restore file in a specific directory and deserializes the datastore
snapshots. The restore file is then deleted.

The DatastoreSnapshotRestore instance needs to be injected into both
DistributedDatastore instances which are created via separate config
system Module instances. However the only way to inject the
DatastoreSnapshotRestore instance would be to define a yang module and
service. I didn't want to go thru the overhead of all that and I didn't
want the DatastoreSnapshotRestore advertised as a service. So I made it
a static singleton that is created via a new bundle Activator class.

The DatastoreSnapshot instance is passed to the ShardManager which
passes each ShardSnapshot to the corresponding Shard actor. On recovery
complete, the RaftActor takes care of applying the restored snapshot.

Change-Id: Ied3db4e49b98320abb34e2acf73b27b29232f8d6
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>