Update patch set 6
Patch Set 6:

> Looks good. I don't see the race condition that you're talking
 > about either.

I talked with TomP after this offline.  Basically the issue is that sometimes the jolokia file is not written prior to the odl-jolokia bundle activation.  I have discovered that odl-jolokia does not properly implement Spring (i.e, it won't honor updates to the file after teh odl-jolokia bundle has loaded).  I have attempted to lessen the likelihood of this condition by doing "configfile", http bundle, then jolokia-osgi bundle.

At worst, you just need to restart karaf in order to properly pick up the config for authentication.  This is not a great approach but it will work until my "authClass" patch is accepted into Jolokia and we can use that to interface the ODL AAA subsystem.

Patch-set: 6