Convert all tests that use ForwardReadyTransaction to use BatchedModifications
Handling ForwardedReadyTransaction is unneccessary. Since it was always a local
message we do not need to worry about having to handle it in an upgrade scenario
as it will not figure in either backwards or forward compatibility.

The ShardTests that cover various scenarios were all using ForwardedReadyTransaction
which I have now migrated to use BatchedModifications. This removes most direct
usages of this message.

Change-Id: I7b3eae972dfa3daf30d24795fc3f34a699f98eab
Signed-off-by: Moiz Raja <moraja@cisco.com>