Update patch set 1
Patch Set 1:

Please hold off on merging this patch until after the code freeze for Lithium-SR3 unless it is a BLOCKING issue.  If this is a BLOCKING issue that requires us to delay the release of Lithium-SR3, please email the release mailing list.  https://lists.opendaylight.org/pipermail/release/2015-November/004518.html

Patch-set: 1