DO NOT MERGE: Bug 4577 Allow specification of a distinct schema cache directory per netconf device
A leaf is added to odl-sal-netconf-connector-cfg.yang to allow specification of
a schema cache directory relative to the "cache" directory.  The leaf defaults to
"schema", so the default directory for loaded yang files is "cache/schema".  It
is useful to specify a distinct cache directory per netconf mount to avoid potential
conflict.

For now, the custom directory requires restart of the controller.  The custom
directory is created but not populated, which leads me to believe that the schema
files are downloaded prior to setting the custom schemaCacheDirectory.  After
restarting the controller, the files download to the correct directory.  Still
working on a solution for this, but not sure how we are going to achieve this, so
please hold off on merging this patch for now.

Change-Id: Ie0c6b653d698c91a443d1b1831306da51bffbf8a
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>