BUG-4626: add dictionary handling
This patch adds dictionary handling for DataTreeCandidatePayload, such
that messages between shard leader and followers reuse dictionaries as
much as possible.

The read dictionary lifecycle is maintained by the NormalizedNodeDataInput,
which throws away the dictionary when the stream indicates it should do
so.

The write dictionary is reset whenever shard leadership changes and this
is communicated to NormalizedNodeDataOutput, which emits the proper
reset/keep instruction next time it is instantiated.

Change-Id: I6a5155f5a55de1755a380a79c76af72f8bfdce42
Signed-off-by: Robert Varga <robert.varga@pantheon.sk>