Update patch set 1
Patch Set 1:

I actually figured you might do this as well :)

The API contract doesn't indicate that unchecked exceptions can be thrown from write/merge/delete so callers don't expect it. But the main reason was that the operations may be performed async in which case the exception wouldn't be propagated to the user at the call site. Prior I think an operation ex caused the transaction to silently succeed on the user end.

Patch-set: 1