Update patch set 18
Patch Set 18:

Given the complexities of persistence and snapshotting wrt the dictionaries, this will need extensive testing. 

I think we should hold off on this work until Boron. I don't feel comfortable merging it this late in Be, particularly since I'm out most of the next 2 weeks and have other tasks to complete so I can't devote much/any time to test this to the extent that it requires. We cannot introduce a latent bug wrt to persistence - it would be catastrophic if data recovery fails in a production environment. I've had to debug recovery issues in a customer deployment in the past and it is difficult and nasty and I'd rather not risk going thru that again.

Patch-set: 18