Update patch set 19
Patch Set 19: Code-Review-2

As I mentioned before, this is too high a risk to try to squeeze in this close to the release, particularly since we really don't even know how much gain we'll get to justify it. I talked to Colin and he agreed. Let's revisit this in Boron when we have more time to really think about it and test it. We have other higher priority bugs/tasks for Be.

Patch-set: 19