Update patch set 9
Patch Set 9:

(1 comment)

I think the code is fine the way it is. Akka guarantees "that the persistent actor will not receive further commands between the persist call and the execution(s) of the associated event handler" so it really doesn't matter when we do the append on tyne success path. 

I assume this patch is really intended for when persistence fails. However in that case, akka stops the actor (by default). 

So I think all we really need to do is remove the FIXME.

Patch-set: 9
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0