Update patch set 1
Patch Set 1:

In addition to this fix we have to make sure the registration of runtime beans gets closed with the module instance to prevent leaking runtime beans.

Or another fix would be to keep just a single runtime bean registrator, as it used to be, modify close method for registration to not close the registrator, just registered beans.

Patch-set: 1