Bug 4948: Catch RuntimeException from ensureParentByMerge
Made the calls to PUT and POST more robust. Caught the RuntimeException from
 the merge call in the EnsureParentsByMerge. If this fails then started
a new transaction and continue with the original PUT. This ensures when a device
is unable to handle the empty list/container merge it will still succeed with the
original request. Similar change has been made for POST.

Change-Id: Icd904b001806dcee2a02eb048ccce4d8d8b8bb9f
Signed-off-by: Balaji Varadaraju <bvaradar@brocade.com>