Update patch set 12
Patch Set 12:

Those tests enable a latch in the InMemoryJournal to block the doAsyncReplyMessages method. The idea is to delay Shard actor startup so it doesn't send the ActorInitialized message to the ShardManager so when findPrimary is sent to the SM it will timeout and return NotInitializedException. I suspect akka isn't calling this method due to the return value from doAsyncReadHighestSequenceNr b/c there's no persisted journal entries setup. Maybe the behavior has changed wrt to the return value of doAsyncReadHighestSequenceNr. If that's the case, adding a persisted journal entry prior to creating the Shard should fix it.

Patch-set: 12