Update patch set 2
Patch Set 2:

DataChangeListeners and DataTreeChangeListeners run on the akka actor threads and can hijack them by doing a blocking activity. This can result in a deadlock happening especially if the DCLs are also blocking on datastore transaction completion.

We anticipated this a long time back and we ensured that we can configure a different dispatcher just for notifications.

If you look at DataChangeListenerProxy and DataTreeChangeListenerProxy where we create the data change actors you will see that we pass a different different dispatcher to these actors. We get this dispatcher from the ActorContext#getNotificationDispatcherPath which either returns the notification dispatcher if it exists or the default dispatcher.

This is the same thing which Muthu said he had to work around but he may not have had to if this configuration was in place.

Patch-set: 2
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0