Bug 4577 Allow specification of a distinct schema cache directory per netconf device
First iteration done here https://git.opendaylight.org/gerrit/#/c/30163
introduce a BUG: sideloaded schemas, manually added in cache/schema, and
specified in the netconf connector, where not being taking in account
in the specified cache folder.

This  patch make sure sideloaded schemas are copied in the specified
directory, and make sure those schemas are correctly processed by
FilesystemSchemaSourceCache.

The reason why we create the directory and add the schemas in there
before create the instance of FilesystemSchemaSourceCache is because
FilesystemSchemaSourceCache loads the schemas from the dir only once,
during the instantiation.

Change-Id: I249e2894f2658300830cdefd40197e169820e8a6
Signed-off-by: adetalhouet <adetalhouet@inocybe.com>