Update patch set 3
Patch Set 3:

I'm wary of the potential overhead of generating a UUID and "toStringing" for every transaction. I think using currentTimeMillis will suffice. The txn ID also includes a counter. So to duplicate an ID across a restart, wall clock time would have to drift back to the exact ms with the same exact counter value. And to cause a problem, a remote shard would have to have stale transaction state for the duplicated ID. I'd say the odds are extremely low.

Patch-set: 3
Label: Verified=0