Make AbstractRaftRPC really abstract
This is an abstract base class, it should be declared as abstract. Also
make the term final, making this an immutable representation. Hide the
class, as it is an internal implementation detail.

Also rework InstallSnapshot to use a serialization proxy, so that the
Serializable (RaftRPC) and Externalizable (InstallSnapshot) is resolved
in a clean way.

Change-Id: Idf0a40470b8ec117df4cfd17d018e18e18f241e8
Signed-off-by: Robert Varga <rovarga@cisco.com>