API Doc Explorer crashes if a yang file on a mounted resource does not have a version
The fix for bug 3782  was checking for dates less than 1970 ( default date when there is no version).
The fix is to check for less than or equal to 1970. The existing code works fine for US time zones
because all US time zones are behind GMT and hence the local time for EPOCH date results in year 1969.
However when in other time zones such as in Europe the year part of the EPOCH date results in 1970.
This is a simple fix to correct that.

Change-Id: I03f9850eaa25a7fbc12895ee0a9eef059a319613
Signed-off-by: Balaji Varadaraju <bvaradar@brocade.com>