Update patch set 13
Patch Set 13:

I finally got 2.4.1 to successfully deploy in karaf. I needed to add a new bundle, akka-protobuf, to the akka-clustering feature. Also 2.4.1 requires com.typesafe.akka/config version 1.3.0. This is defined in odlparent so I'll need to push a patch for that.

Also, 2.4.1 no longer provides leveldb as the default journal in their reference.conf so we need to specify it in our akka.conf. However, to ease upgrades, I want to wait until https://git.opendaylight.org/gerrit/#/c/34553/ is merged so we can specify it in the factory akka.conf.

Patch-set: 13
Label: Verified=0