Bug 5485: Improve DataTreeModification pruning on recovery
Modified the PruningDataTreeModification and NormalizedNodePruner to
validate path and node QNames via the SchemaContext instead of just
namespaces. This allows migration support for any element to be removed
from a yang hierarchy.

Also handled SchemaValidationFailedException on ready which can happen
with writes which don't immediately validate the sctructure as merge
does. The modification tree is re-applied with pruning.

Change-Id: Ic355de40d43b9c2075d54004776c10866552cfd1
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>