Convert toaster-it to use mdsal-it
I recently added an mdsal dependency to naother project and it caused
the toaster-it test to fail. There was no useful output to troubleshoot,
just that some bundles failed to resolve.

It turns out the mdsal dependency and all its dependencies would need to be
added to the bundle configuration for the test. This will be ongoing issue
every time a new dependency is added that might affect the toaster test.

To alleviate this I converted the test from naually installing bundles to use
the mdsal-it functionality which installs features into the
opendaylight-karaf-empty distro.

I also changed AbstractConfigTestBase to override the pax logging to
append to stdout, otherwise most output goes to karaf.log. This is helpful to
troubleshoot failures.

For some reason the ToasterTest fails in Jenkins builds due to failure to find
the features-mdsal artifact via the maven URL. I tried to copy the
artifact locally via the maven-dependency-plugin and referencing the
local file, which worked for the features-mdsal artifact but not the
transitive dependencies in the feature.xml. Therefore I confgured the
test not to fail the build. But at least the ToasterTest is up-to-date
and no longer uses a deprecated API so it's still useful as an example.

Change-Id: I1dd9150c6e617e491458560fe4b360d37548f264
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>