Update patch set 5
Patch Set 5:

No, I meant to fix that in all branches. Although I'm not using this functionality so I don't really mind the fix not being in lithium...

Sad to hear there will be no more SR for lithium as I have a couple of fix I need to come up with for the lithium release.

I'll let the patch be in lithium. AFAIK it can still be merged, and the .5-SNAPSHOT artifacts will still be generated.

Patch-set: 5