Update patch set 5
Patch Set 5:

I assume you're referring to my comment. That section of code referred to was patched into stable/be to workaround an issue - https://git.opendaylight.org/gerrit/#/c/36172/.  But it's not an ideal solution (bordering on hack) and I'm planning a better solution in master to be done with https://git.opendaylight.org/gerrit/#/c/36171/. 

One of the unit tests for this patch indirectly causes the issue to occur so that's why I had added that code here originally to work around it. So this patch is actually ready to go as is and can be merged. As I said I'll remove the code in question with a cleaner solution in a follow-up.

Patch-set: 5
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0