Add restart-dependents-on-update blueprint extension
Added an attribute extension to the <blueprint> element that adds a bean
processor that scans for any "property-placeholder" elements reacts to
changes to the corresponding cfg file by restarting this blueprint
container any dependent containers that consume OSGi services provided by
this container in an atomic and orderly manner.

A new OSGi service, BlueprintContainerRestartService, was added that
performs the orderly restart. This service is created and registered by
the BlueprintBundleTracker.

Change-Id: I5f463303c3a8aba35b3ed914268bdc67ac795a5a
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>