Fix ShardManagerInfo JMX registration error
The ShardManagerInfo mbean is failing to register with error:
"ShardManagerInfo does not implement DynamicMBean, and neither follows
the Standard MBean conventions". To be a standard MBean, the
implementation class must be in the same package as the MBean interface.
However recent refactoring moved the ShardManagerInfo class to a new
package so I moved both the ShardManagerInfo and ShardManagerMBean
classes to a new package
org.opendaylight.controller.cluster.datastore.shardmanager.jmx.

(javax.management.NotCompliantMBeanException: Class
org.opendaylight.controller.cluster.datastore.shardmanager.ShardManagerInfo
is not a JMX compliant Standard MBean) nor the MXBean conventions"

Change-Id: I77870e7cf8274f3e99e116d6f18949ac3cf26e41
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>