Update patch set 5
Patch Set 5:

yes, it is exist. but we have to provides different ways to show the info. like in development I'm not using jconsle, just need to show the number of transactions using karaf CLI. it is really helpful instead of setup jconsole client to connect to karaf and see the info.

Patch-set: 5