Update patch set 3
Patch Set 3:

I tried the switch shard state operation and I entered the shard name "default" but it expects the full shard id "member-1-shard-default-config". That caused an NPE in ShardManager which was propagated to akka and it restarted the ShardManager but the Shards weren't recreated. That's not good - potential DOS attack. I think we should validate the shard id in ShardManagerInfo and pass the ShardIdentifier to the ShardManager.

Patch-set: 3
Label: Verified=0