Update patch set 2
Patch Set 2:

replyTo was added for a potential future enhancement to allow for a user-initiated leadership transfer.  In that case we would want to reply back to the caller. But if we ever add that feature, replyTo could be local to the OnComplete callback and need not be passed to  the cohort.

Patch-set: 2
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0