Update patch set 1
Patch Set 1:

By commit going through I mean the commitIndex being incremented. The state is applied after the commitIndex gets incremented. So let's say I had a bunch of append entries / replies in the queue I would rapidly increment commitIndex  and then schedule an ApplyState on the leader so if the queue was really backed up then it would take time before the state was applied.

I agree with you that for this logging itself does not seem to add a lot of value - I haven't found it particularly interesting. How about we just remove it altogether?

Patch-set: 1