Update patch set 9
Patch Set 9:

Why just shard-topology-operational and not every shard? It seems you're tailoring the output to your needs and what you want to see which makes it less generally useful.  It's likely you will be the only user of this so why don't you just keep it internal instead of pushing upstream? Then you can do whatever you want with it.

Patch-set: 9