Update patch set 23
Patch Set 23:

Yes - correct. I implemented the user config via ConfigAdmin. This part is up for debate and discussion and was discussed on the call. The netty config knobs would not commonly be tweaked by end users and would likely require intimate engineering knowledge to do so so I don't think it really matters if they're configurable via netconf or not. However we could honor the CSS settings and updates or use the config datastore although I'm not sure if the netty bundles can depend on the datastore.

As it stands now they're configurable via ConfigAdmin - we can change this part later if need be. Right now I'd like to get these changes in so apps can start using blueprint.

Patch-set: 23
Label: Verified=0