Update patch set 1
Patch Set 1:

> What are the offender and why are we not fixing them properly?

There's actually quite a few and I came to realize that this was going to be a very significant effort when I made this patch. My main problem was all the {#link method()} in Controller that pointed to invalid references. I don't know the code base well enough to know where to point these references to or if they are still valid as in some cases entire methods were no longer found in the project.

Patch-set: 1