BUG-5280: add backend handlers for new messages
Adds the basic request routing infrastructure inside the Shard. This
tracks frontend generations and also mapping to handler actors.

It introduces LocalHistoryActor, which is a backend tracking status
of local histories as created by the frontend. Once a transaction is
completed, the request to persist it is sent to the Shard, which is
responsible for applying the transaction to the global history and
notifying the appropriate frontend and the local history actor.

Change-Id: Ie2cdbff940741cee2d68df4948b8fdcfe91d34e5
Signed-off-by: Robert Varga <rovarga@cisco.com>